+++ build.gradle
@@ -31,10 +31,6 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client' // OAuth
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
-	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 
 	// 2. 개발 편의 도구
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
