+++ docker-compose.yml
@@ -11,8 +11,6 @@ services:
     depends_on:
       db:
         condition: service_healthy
-      redis: # <-- 추가
-        condition: service_started
     restart: unless-stopped
 
   db:
