+++ src/main/java/dripnote/home/payload/response/HomeResponse.java
@@ -1,12 +1,10 @@
 package dripnote.home.payload.response;
 
-import lombok.Builder;
 import dripnote.home.payload.dto.HomeBeanDTO;
 import dripnote.home.payload.dto.HomeTastingsDTO;
 
 import java.util.List;
 
-@Builder
 public record HomeResponse(
         List<HomeTastingsDTO> tastings,
         List<HomeBeanDTO> beans
