--- src/main/java/dripnote/security/payload/dto/TokenResponse.java
@@ -1,11 +1,0 @@
-package dripnote.security.payload.dto;
-
-import lombok.Builder;
-
-@Builder
-public record TokenResponse(
-    String accessToken,
-    String refreshToken,
-    String tokenType // "Bearer"
-) {
-}
