+++ src/main/java/dripnote/user/domain/User.java
@@ -1,28 +1,23 @@
 package dripnote.user.domain;
 
-import dripnote.common.domain.BaseTimeEntity;
 import dripnote.user.enums.UserProvider;
 import dripnote.user.enums.UserRole;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 
 import java.time.LocalDateTime;
 
 @Getter
-@Builder
+@Setter
 @NoArgsConstructor
-@AllArgsConstructor
 @Entity
 @Table(name = "users", uniqueConstraints = {
-        @UniqueConstraint(name = "unique_provider_id", columnNames = {"provider", "provider_id"}),
-        @UniqueConstraint(name = "unique_nickname", columnNames = {"nickname"})
+        @UniqueConstraint(name = "unique_provider_id", columnNames = {"provider", "provider_id"})
 })
-public class User extends BaseTimeEntity {
+public class User {
     /**
      * updatedAt, profileImageUrl, status -> 해당 데이터는 현재 불필요할 것 같아서 삭제했습니다!
      */
@@ -31,7 +26,7 @@ public class User extends BaseTimeEntity {
     @Column(name = "user_id")
     private Long userId;
 
-    // 카카오는 null일 수 있습니다.
+    // OAuth 정책상 이메일은 선택인 경우가 많아 nullable = true(디폴트)로 변경
     @Column(name = "email", length = 255)
     private String email;
 
@@ -57,11 +52,8 @@ public class User extends BaseTimeEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "role", nullable = false, length = 20)
     private UserRole role = UserRole.USER;
-    /**
-     * 닉네임 업데이트 (검증은 Service에서 수행)
-     * 이 메서드는 이미 검증된 닉네임만 받아서 업데이트합니다.
-     */
-    public void updateNickname(String newNickname) {
-        this.nickname = newNickname;
-    }
+
+    @CreationTimestamp
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt;
 }
