--- src/main/java/dripnote/user/payload/dto/UserUpdateRequest.java
@@ -1,6 +1,0 @@
-package dripnote.user.payload.dto;
-
-public record UserUpdateRequest(
-        String nickname
-) {
-}
